backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
crewai==0.80.0
crewai-tools==0.14.0
google-generativeai==0.8.3
langchain==0.3.7
langchain-google-genai==2.0.6
langchain-core==0.3.19
beautifulsoup4==4.12.3
requests==2.32.3
httpx==0.27.2
jinja2==3.1.4
aiosqlite==0.20.0